Renamed org.scalatest.check to org.scalatestplus.scalacheck.
