+++ scalatest-test/src/test/scala/org/scalatestplus/scalacheck/PropertyCheckConfigHelperSuite.scala
@@ -13,36 +13,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
-package check
+package org.scalatestplus
+package scalacheck
 
 import org.scalactic.anyvals._
+import org.scalatest._
 
-class PropertyCheckConfigurationHelperSuite extends FunSuite with Matchers {
+@deprecated("Remove when removing PropertyCheckConfig")
+class PropertyCheckConfigHelperSuite extends FunSuite with Matchers {
 
-  import org.scalatest.check.ScalaCheckConfiguration._
+  import ScalaCheckConfiguration._
 
   val DefaultMinSuccessful: PosInt = 9
   val PassedMinSuccessful: PosInt = 3
 
+  val DefaultMaxDiscarded = 99
+  val PassedMaxDiscarded = 33
+
   val DefaultMinSize: PosZInt = 99
   val PassedMinSize: PosZInt = 33
+
+  val DefaultMaxSize = 99
+  val PassedMaxSize = 33
 
   val DefaultWorkers: PosInt = 99
   val PassedWorkers: PosInt = 33
 
-  val DefaultSizeRange: PosZInt = 0
-  val PassedSizeRange: PosZInt = 10
-
-  val DefaultMaxDiscardedFactor: PosZDouble = 1.0
-  val PassedMaxDiscardedFactor: PosZDouble = 0.5
-
   val defaultConfig =
-    PropertyCheckConfiguration(
+    PropertyCheckConfig(
       minSuccessful = DefaultMinSuccessful,
-      maxDiscardedFactor = DefaultMaxDiscardedFactor,
+      maxDiscarded = DefaultMaxDiscarded,
       minSize = DefaultMinSize,
-      sizeRange = DefaultSizeRange,
+      maxSize = DefaultMaxSize,
       workers = DefaultWorkers
     )
 
@@ -59,7 +61,7 @@ class PropertyCheckConfigurationHelperSuite extends FunSuite with Matchers {
   }
 
   test("getParams returns default minSuccessful config param if none passed") {
-    val params = getParams(Seq(Workers(DefaultWorkers)), defaultConfig)
+    val params = getParams(Seq(MaxDiscarded(PassedMaxDiscarded)), defaultConfig)
     params.minSuccessfulTests should equal (DefaultMinSuccessful.value)
   }
 
@@ -68,19 +70,19 @@ class PropertyCheckConfigurationHelperSuite extends FunSuite with Matchers {
 
   // maxDiscarded
   test("getParams returns passed maxDiscarded config param") {
-    val params = getParams(Seq(MaxDiscardedFactor(PassedMaxDiscardedFactor)), defaultConfig)
-    params.maxDiscardRatio should equal (PassedMaxDiscardedFactor.value)
+    val params = getParams(Seq(MaxDiscarded(PassedMaxDiscarded)), defaultConfig)
+    params.maxDiscardRatio should equal (maxDiscardRatio(PassedMaxDiscarded + 1, params.minSuccessfulTests))
   }
 
   test("getParams throws IAE if passed multiple maxDiscarded config params") {
     intercept[IllegalArgumentException] {
-      getParams(Seq(MaxDiscardedFactor(33.0), MaxDiscardedFactor(34.0)), defaultConfig)
+      getParams(Seq(MaxDiscarded(33), MaxDiscarded(34)), defaultConfig)
     }
   }
 
   test("getParams returns default maxDiscarded config param if none passed") {
     val params = getParams(Seq(MinSuccessful(PassedMinSuccessful)), defaultConfig)
-    params.maxDiscardRatio should equal (DefaultMaxDiscardedFactor.value)
+    params.maxDiscardRatio should equal (maxDiscardRatio(DefaultMaxDiscarded + 1, params.minSuccessfulTests))
   }
 
   // minSize
@@ -100,32 +102,28 @@ class PropertyCheckConfigurationHelperSuite extends FunSuite with Matchers {
     params.minSize should equal (DefaultMinSize.value)
   }
 
-  // sizeRange
-  test("getParams returns passed sizeRange config param") {
-    val params = getParams(Seq(SizeRange(PassedSizeRange)), defaultConfig)
-    params.maxSize should equal (DefaultMinSize + PassedSizeRange)
-  }
-
-  test("getParams returns passed minSize and sizeRange config param") {
-    val params = getParams(Seq(MinSize(PassedMinSize), SizeRange(PassedSizeRange)), defaultConfig)
-    params.maxSize should equal (PassedMinSize + PassedSizeRange)
+  // maxSize
+  test("getParams returns passed maxSize config param") {
+    val params = getParams(Seq(MaxSize(PassedMaxSize)), defaultConfig)
+    params.maxSize should equal (PassedMaxSize)
   }
 
   test("getParams throws IAE if passed multiple maxSize config params") {
     intercept[IllegalArgumentException] {
       getParams(Seq(MaxSize(33), MaxSize(34)), defaultConfig)
     }
-    intercept[IllegalArgumentException] {
-      getParams(Seq(MaxSize(33), SizeRange(34)), defaultConfig)
-    }
-    intercept[IllegalArgumentException] {
-      getParams(Seq(SizeRange(33), SizeRange(34)), defaultConfig)
-    }
   }
 
-  test("getParams returns default sizeRange config if none passed") {
+  test("getParams returns default maxSize config param if none passed") {
     val params = getParams(Seq(MinSuccessful(PassedMinSuccessful)), defaultConfig)
-    params.maxSize should equal (DefaultMinSize + DefaultSizeRange)
+    params.maxSize should equal (DefaultMaxSize)
+  }
+
+  test("getParams returns default maxSize config param if none passed and MinSuccessful changed") {
+    val params = getParams(Seq(MinSize(PassedMinSize)), defaultConfig)
+    println("Params.maxsize: " + params.maxSize)
+    params.maxSize should equal (DefaultMaxSize)
+    params.minSize should equal (PassedMinSize.value)
   }
 
   // workers
@@ -148,19 +146,18 @@ class PropertyCheckConfigurationHelperSuite extends FunSuite with Matchers {
   test("getParams returns all default if no config params passed") {
     val params = getParams(Seq(), defaultConfig)
     params.minSuccessfulTests should equal (DefaultMinSuccessful.value)
-    params.maxDiscardRatio should equal (DefaultMaxDiscardedFactor.value)
+    params.maxDiscardRatio should equal (maxDiscardRatio(DefaultMaxDiscarded + 1, params.minSuccessfulTests))
     params.minSize should equal (DefaultMinSize.value)
-    params.maxSize should equal (DefaultMinSize.value + DefaultSizeRange.value)
+    params.maxSize should equal (DefaultMaxSize)
     params.workers should equal (DefaultWorkers.value)
   }
 
   test("getParams returns all passed if all config params passed") {
-    val params = getParams(Seq(MinSuccessful(PassedMinSuccessful), MaxDiscardedFactor(PassedMaxDiscardedFactor), MinSize(PassedMinSize),
-      SizeRange(PassedSizeRange), Workers(PassedWorkers)), defaultConfig)
+    val params = getParams(Seq(MinSuccessful(PassedMinSuccessful), MaxDiscarded(PassedMaxDiscarded), MinSize(PassedMinSize), MaxSize(PassedMaxSize), Workers(PassedWorkers)), defaultConfig)
     params.minSuccessfulTests should equal (PassedMinSuccessful.value)
-    params.maxDiscardRatio should equal (PassedMaxDiscardedFactor.value)
+    params.maxDiscardRatio should equal (maxDiscardRatio(PassedMaxDiscarded + 1, params.minSuccessfulTests))
     params.minSize should equal (PassedMinSize.value)
-    params.maxSize should equal (PassedMinSize.value + PassedSizeRange.value)
+    params.maxSize should equal (PassedMaxSize)
     params.workers should equal (PassedWorkers.value)
   }
 }
