+++ scalatest/src/main/scala/org/scalatest/prop/package.scala
@@ -20,22 +20,22 @@ package org.scalatest
  */
 package object prop {
 
-  @deprecated("Please use org.scalatest.check.Checkers instead.")
-  type Checkers = org.scalatest.check.Checkers
+  @deprecated("Checkers has been moved from org.scalatest.prop to org.scalatestplus.scalacheck. Please update your imports, as this deprecated type alias will be removed in a future version of ScalaTest.")
+  type Checkers = org.scalatestplus.scalacheck.Checkers
 
-  @deprecated("Please use org.scalatest.check.Checkers instead.")
-  lazy val Checkers = org.scalatest.check.Checkers
+  @deprecated("Checkers has been moved from org.scalatest.prop to org.scalatestplus.scalacheck. Please update your imports, as this deprecated type alias will be removed in a future version of ScalaTest.")
+  lazy val Checkers = org.scalatestplus.scalacheck.Checkers
 
-  @deprecated("Please use org.scalatest.check.ScalaCheckDrivenPropertyChecks instead.")
-  type GeneratorDrivenPropertyChecks = org.scalatest.check.ScalaCheckDrivenPropertyChecks
+  @deprecated("GeneratorDrivenPropertyChecks has been moved from org.scalatest.prop to org.scalatestplus.scalacheck and renamed as ScalaCheckDrivenPropertyChecks. Please update your imports, as this deprecated type alias will be removed in a future version of ScalaTest.")
+  type GeneratorDrivenPropertyChecks = org.scalatestplus.scalacheck.ScalaCheckDrivenPropertyChecks
 
-  @deprecated("Please use org.scalatest.check.ScalaCheckDrivenPropertyChecks instead.")
-  lazy val GeneratorDrivenPropertyChecks = org.scalatest.check.ScalaCheckDrivenPropertyChecks
+  @deprecated("GeneratorDrivenPropertyChecks has been moved from org.scalatest.prop to org.scalatestplus.scalacheck and renamed as ScalaCheckDrivenPropertyChecks. Please update your imports, as this deprecated type alias will be removed in a future version of ScalaTest.")
+  lazy val GeneratorDrivenPropertyChecks = org.scalatestplus.scalacheck.ScalaCheckDrivenPropertyChecks
 
-  @deprecated("Please use org.scalatest.check.ScalaCheckPropertyChecks instead.")
-  type PropertyChecks = org.scalatest.check.ScalaCheckPropertyChecks
+  @deprecated("PropertyChecks has been moved from org.scalatest.prop to org.scalatestplus.scalacheck and renamed as ScalaCheckPropertyChecks. Please update your imports, as this deprecated type alias will be removed in a future version of ScalaTest.")
+  type PropertyChecks = org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 
-  @deprecated("Please use org.scalatest.check.ScalaCheckPropertyChecks instead.")
-  lazy val PropertyChecks = org.scalatest.check.ScalaCheckPropertyChecks
+  @deprecated("PropertyChecks has been moved from org.scalatest.prop to org.scalatestplus.scalacheck and renamed as ScalaCheckPropertyChecks. Please update your imports, as this deprecated type alias will be removed in a future version of ScalaTest.")
+  lazy val PropertyChecks = org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 
 }
