+++ scalatest/src/main/scala/org/scalatestplus/scalacheck/ScalaCheckPropertyChecks.scala
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
-package check
+package org.scalatestplus
+package scalacheck
 
 import org.scalatest.prop.TableDrivenPropertyChecks
